Create hex_assignments
